Deal with GOPATH and GOPATH/bin for golang
Okay, so we always set `GOROOT` variable to
`~/.eclectica/go/current` that's a given.

Now about `GOPATH` –
go versions lower then 1.7 do not have default `GOPATH` environment variable.
Starting from 1.7 `GOPATH` is now set to `~/go` path
(see go help gopath for more). We do the same if for other
versions as a default, but only if user didn't set themselves.

As with `GOPATH/bin` we just set it when first command is run in
to some rc file.

Fixes #14